+++ features/step_definitions/shop/admin/shop_admin_offers_steps.rb
@@ -28,37 +28,22 @@ Given(/^I submit a valid new offer$/) do
   @ShopAdminOffers.select_store
   @ShopAdminOffers.click_create_sidebar_button
 end
+
 Given(/^I verify the offer has been created$/) do
   expect(page).to have_css(".alert-success", :text =>"Offer successfully created.")
+  @offer_url = page.current_url
 end
-
-Given(/^I delete the offer$/) do
-  step 'I visit "ShopAdminOffers"'
-  @ShopAdminOffers.filter_datatable_for_offer
-  @ShopAdminOffers.click_on_offer
-  @ShopAdminOffers.delete_offer
-end
-
-Given(/^I verify the offer has been deleted$/) do
-  expect(page).to have_css(".alert-success", :text => "Offer successfully destroyed")
-end
-
 
 Given(/^I add the new category to the offer$/) do
-  @ShopAdminOffers.search_for_offer_category
+  @ShopAdminOffers.search_for_offer_category(@category_label)
   @ShopAdminOffers.select_category_result_return
   @ShopAdminOffers.update_offer_page
 end
 
 Given(/^I verify the category has been added to the offer$/) do
-  expect(page).to have_css("#offer_categories_attributes_0_label[value='Fitness & Diet']")
+  expect(page).to have_css("#offer_categories_attributes_0_label[value='#{@category}']")
 end
 
 Given(/^I verify the offer has been updated$/) do
-  expect(page).to have_css(".alert-success", :text => "× Offer successfully updated.")
-end
-
-Given(/^I delete the offer category$/) do
-  step 'I visit "ShopAdminCategories"'
-  @ShopAdminCategories.delete_test_category
+  expect(page).to have_css(".alert-success", :text => "Offer successfully updated.")
 end