+++ features/support/test_helper.rb
@@ -132,9 +132,15 @@ module HelperMethods
     @admin_users.visit
   end
 
-  def visit_shop_categories_in_new_window
+  def visit_shop_admin_resource_datatable_in_new_window(resource)
     @ShopAdminCategories = ShopAdminCategories.new
     @ShopAdminCategories.login_in_new_window
+    visit(resource)
+  end
+
+  def delete_shop_admin_resource
+    click_link "Delete"
+    page.driver.browser.switch_to.alert.accept
   end
 
   def verify_discount(original_product_amt_string, actual_product_discounted_amt_string, discount_percentage, options={})